Move performance fixture helper into shared workflow utilities
- moved modules/performance/fixture-roots.js into shared/workflow/fixture-roots.js so workflow path helpers stay alongside related utilities used across the CLI

- updated the performance module and tests to read from the new shared location and resolve bundled metadata paths via REPO_ROOT
