+++ src/cli/test/feather-metadata.test.js
@@ -6,9 +6,14 @@ import {
     FEATHER_METADATA_PATH,
     loadBundledFeatherMetadata
 } from "../src/modules/feather/metadata.js";
+import { REPO_ROOT } from "../src/shared/workspace-paths.js";
 
 test("feather metadata loader resolves bundled snapshot", () => {
-    const expectedPath = path.resolve("resources", "feather-metadata.json");
+    const expectedPath = path.resolve(
+        REPO_ROOT,
+        "resources",
+        "feather-metadata.json"
+    );
     assert.equal(FEATHER_METADATA_PATH, expectedPath);
 
     const metadata = loadBundledFeatherMetadata();
